requirements.txt
playwright>=1.52.0
PyYAML>=6.0.2
requests>=2.32.3
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
pypdf>=5.0.0
python-docx>=1.1.2
supabase>=2.27.0
stripe>=11.0.0
python-dotenv>=1.0.1
